refactor: add game folder with common game components
