+++ my-app/src/components/game/LevelButton.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { ModalProps } from './LevelModal';
 
 interface LevelButtonProps {
