+++ my-app/src/components/game/Timer.tsx
@@ -1,4 +1,4 @@
-import React, { Dispatch, SetStateAction } from 'react'
+import { Dispatch, SetStateAction } from 'react'
 
 interface TimerProps {
   time: number;
